Address review feedback on the Activity wrapper
- Add deprioritize and deprioritized to the project dictionary, so the
  spellcheck job passes.
- Merge ScreenActivityWrapper/index.native.tsx into index.tsx. The two
  files were identical apart from their comments, because
  CustomViewWrapper already resolves per platform.
- Share one NonTopScreenWrapperProps between the wrapper and the
  descriptor mapping instead of declaring the same shape twice.
- Correct the isScreenBlurred doc. A covered screen often stays visible,
  which is the whole point of CustomViewWrapper.
- Write display contents on the element even where MutationObserver is
  missing, and key the guard on the element so a new node is served too.
- Say in useIsScreenCovered where each half of the covered state comes
  from, and note that the wrapper reads the navigator's focus.
- Split hyphenated clauses into sentences, and state that the native view
  config was validated on Fabric.
