+++ cspell.json
@@ -561,6 +561,8 @@
         "delish",
         "dependentaxis",
         "deployers",
+        "deprioritize",
+        "deprioritized",
         "deprioritizes",
         "describedby",
         "devportal",
