--- src/libs/Navigation/PlatformStackNavigation/createPlatformStackNavigatorComponent/ScreenActivityWrapper/index.native.tsx
@@ -1,42 +1,0 @@
-import useThemeStyles from '@hooks/useThemeStyles';
-
-import CustomViewWrapper from '@libs/Navigation/PlatformStackNavigation/createPlatformStackNavigatorComponent/CustomViewWrapper';
-
-import React, {Activity} from 'react';
-
-import type ScreenActivityWrapperProps from './types';
-
-import useIsScreenCovered from './useIsScreenCovered';
-import useScreenActivityMode from './useScreenActivityMode';
-
-/**
- * Deprioritizes rendering of a blurred screen with React <Activity>. Unlike react-freeze, a hidden Activity keeps
- * processing state updates at background priority and runs effect cleanups when hiding. A modal or popover that is
- * still dismissing when the screen gets blurred always finishes its close chain, so no freeze delay and no modal
- * coordination are needed here. useScreenActivityMode decides when the screen is deprioritized.
- *
- * The painted content is stale for as long as the screen is covered, so it is marked inert for that whole time,
- * which keeps it out of the accessibility tree and out of the reach of touches. The flag follows the navigation
- * state rather than the Activity mode, because a reveal is deferred until the transition ends and the screen the
- * user is already looking at has to be usable right away.
- */
-function ScreenActivityWrapper({isScreenBlurred, routeKey, routeName, children}: ScreenActivityWrapperProps) {
-    const styles = useThemeStyles();
-    const isScreenCovered = useIsScreenCovered(isScreenBlurred);
-    const mode = useScreenActivityMode({isScreenCovered, routeKey, routeName});
-
-    // A hidden Activity hides its content through the same mechanism as a suspended tree, so CustomViewWrapper
-    // keeps the content painted and the underlay screen stays visible during swipe-back gestures.
-    return (
-        <Activity mode={mode}>
-            <CustomViewWrapper
-                style={styles.flex1}
-                inert={isScreenCovered}
-            >
-                {children}
-            </CustomViewWrapper>
-        </Activity>
-    );
-}
-
-export default ScreenActivityWrapper;
